Clear m_fileHandle on close.
This resolves a potential case of undefined behavior. Consider the following code

{
    LC3Writer writer("file.obj");

    writer.close();
}

Previous to this change, LC3Writer::close() will leave the value of m_fileHandle
as it is, which in this case is not equal to nullptr. When writer's destructor
is called, LC3Writer::close() is called once again, but this time on a freed
pointer. Thus, the above code invoked undefined behavior.