Reimplement StringView's iterator class.
I originally wrote the previous iterator class with the intention
that I would write a more complete version at a later time. This
is that more complete version. It has random-access capabilities, and
dereferencing the end iterator returns a null character. Spiffy!

And because of its size, I decided to place its definition in a separate
file.