Change StringViewIterator::operator*()'s return type.
Before, the return value was a non-reference value, meaning that one cannot take
the address of this method's return value. This may be useful in some situations,
so to allow for this to take place, I changed the return type to an lvalue reference.
Something special (i.e., a hack) had to be done in order to accommodate the end
iterator dereferencing to a null character, but it's nothing too bad.