Add the TemplateString macro
TemplateString is a preprocessor macro that defines a thin wrapper class
that enables one to pass string values as template parameters. There are
two parameters to the TemplateString macro, a name parameter which defines
the name of the wrapper class, and the string value itself.

I defined TemplateString in the GenericParser.cpp file because I really
only intended to use it in conjunction with this class. I may move its
definition to another file in the future.