Use an alternate method of representing file modes.
The previous way of representing file modes were simple classes
with an open() function that created a file handle in the respective mode.
Not only was this method rather limiting in terms of functionality, but it was
also a potential source of bugs. Any class with a static open() function
could be used as a template parameter. So as an exercise I wanted to
design this system better, and I think I did it.

By defining the FileModeTraits class, I can not only validate template
parameters, but I can also retrieve information on these parameters
all in one handy class.