Pass stored savepoint by rvalue reference in destructor
In the destructor to SavepointGuard, it can be guaranteed that the stored
savepoint will not be used again. Therefore, it is safe to perform a move
of this value. By passing the savepoint to Context::restore() via rvalue
reference, the context implementation is able to perform move semantics,
possibly making the savepoint restoration process more efficient.