Change behavior of ParserElementTraits::IsCompatible
ParserElementTraits::IsCompatible's parameter is now a parser class as
opposed to a parser context class. The reasoning behind this is that
parser elements are just that: elements of a parser. As such, they are
more closely associated with them, not the respective context class.
Therefore, to reflect that, I am making this change.