Use SFINAE to compute ParserElementTraits::IsCompatible
Suppose a ParserElementTraits template is instantiated with a non-valid
parser element and IsCompatible is computed. Suppose this invalid class
has an invalid (even non-existant) parse function. IsCompatible directly
calls decltype() to determine whether or not a parse function exists and
has the correct function signature. However, because the invalid class
has an invalid parse function, the decltype will fail to deduce a type,
resulting in the compiler throwing an error. What is desired is for
IsCompatible to simply evaluate to false in this case.
    The desired behavior can be achieved with SFINAE, as type deduction
failures are explicitly allowed under this system. A simple helper
class is introduced to implement the SFINAE rules.