Migrate all utility classes to the 'util' directory
For the sake of better code organization, I migrated all classes
that were not directly related to LC-3 to the utility directory.

The decision to squash this merge was made because many of the constituent
commits included intermediate changes which were never intended to be in
the final result. These intermediate changes introduce noise into diffs,
and so in an effort to reduce noise I am only showing the final result.

Squashed commit of the following:

commit 2ca8c74cae577e44198512f1a815fdac0623809a
Author: Micah Duron <micah@duron.me>
Date:   Wed May 10 20:35:02 2017 -0700

    Move StringTokenizer to the 'util' directory

commit d0eed794326b11e7ffafd38dc0b14fd786248024
Author: Micah Duron <micah@duron.me>
Date:   Wed May 10 20:24:51 2017 -0700

    Move StringView to the 'util' directory

commit ac8a98befed0cce3f75edd4a8bf9589b28a0c5bb
Author: Micah Duron <micah@duron.me>
Date:   Wed May 10 20:04:45 2017 -0700

    Move CharClass to the 'util' directory