Fix logic error in Util::GenericParser::Multiple
This change is two-fold. The first change concerns the logical-OR.
I apprently wasn't thinking when I originally wrote this, because
this function would return true if instanceCount met the
lower-bound requirements, regardless of the upper-bound. Changing
the logical-OR to a logical-AND resolves this.
The second change deals with the ternary expression. Apparently
the syntax mandates that parentheses be put around a nested ternary
expression, because if not placed there, the entire expression is
incorrect.