Change return type of dereference operator overload
Because of the special nature of this iterator (the terminating
iterator always dereferences to a null character), having the
dereference operator return a reference complicated the code.
After some testing, I determined that it was safe enough to
change the return type of the dereference operator from a
reference to a temporary. This move significantly simplified
the code.