Rewrite conditional for IsCompatible_Helper
The goal of IsCompatible_Helper is to verify that an element has a
static parse() function exists, it has the proper parameters,
and that it returns a bool. Using std::void_t to accomplish all of
this is possible, but it requires a hack. Fortunately, std::enable_if
used in conjunction with std::is_same can achieve the intended
goal without any hackery.