Add the project's root directory to the default include list
This makes it possible to write code that is location agnostic.
Consider a project with the subdirectories util/ and tests/.
Before this commit, code in tests/ including from util/ would
require an include statement in the form
    #include "../util/SomeHeader.h" .
If I were to move the code in tests/ to util/tests/, that statement
would need to be edited to reflect that move.

By adding the root directory to the default include list, the above
include statement could be changed to
    #include <util/SomeHeader.h>
and it would function correctly regardless of its location relative
to util/.