Reimplement CharClass
The original implementation of CharClass used callable objects to
classify characters. There are some limitations to this method
as I will explain below.
    Consider you want to define a class that returns true only
for the characters 'a' and 'b', and false everywhere else. A
typical definition looks like
    CharClass someName( [](char c) { return c == 'a' || c == 'b' } ); .
Not very elegant as you can see. And if more characters are added,
the more messy things become. Furthermore, this object cannot be
declared constexpr due to std::function not being constexpr,
which is used internally in the old implementation.
    The solution presented here resolves this by replacing the
callable with a std::bitset object that acts like a lookup table
to the entire character range. A character can be used to index
into this bit set and the returned value is the class to which
that character belongs.
    This method dramatically simplifies the above definition
with the introduction of a new constructor that accepts a string
containing the characters belonging to the class.
    CharClass someName("ab");
The string will then be used to populate the bit set. Additionally,
this method is backwards compatible with the previous implementation,
in that CharClass still provides a constructor that accepts callable
objects. The only difference is that non-pure functions will not
behave the same under this new implementation (non-pure classification
functions were never a good idea to begin with).
    The new implementation is constexpr compatible and potentially
more efficient since classification is now a table lookup instead
of an arbitrary function call.
    The main disadvantage of this new implementation is the
increased memory overhead. Because the bit set must span the whole
character range (256), the amount of memory a single CharClass object
consumes is at least 32 bytes (one bit per character). Nevertheless,
I think that the pros outweigh the cons. The amount of increased
memory usage is justified considering the amount of benefits that
are gained.