Declare static class members as inline
The previous code would have thrown an error had I included the header
Constants.h in multiple source files due to the one-definition rule.
This is because Constants::Directives and Constants::Instructions
had their definition in the header itself. Previous C++ standards
did not permit this, but C++17 does if the members are declared
inline.