Add new method of error handling in parser
These new meta parser elements temporarily change the value of the
parser flag corresponding to the error mode. These elements can be
used to establish 'blocks' in which the error mode takes on a certain
value. Elements executed within these blocks can query the error mode
and can modify their error handling behavior accordingly.
    An example of where such a mechanism might be useful is in the
parsing of decimal numbers. In LC-3, a decimal number is any string
that satisfies the following regex: '#?-?[0-9]+'. There is a potential
ambiguity here, as the grammar rules for defining a label name allow
for a numeric character to be the leading character. As such, the token
'10' may be interpreted as either a label or a decimal number. Thus,
writing a grammar rule for parsing this string will look something
like this:
    Either<DecimalNumber, LabelName>;
This can be read as 'either parse the token as a decimal number, or if
that fails, parse it as a label name.' In the case of the token '10',
this will clearly be parsed as a decimal number. But consider the token
'1az'.
    The above grammar rule will first attempt to parse this with the
DecimalNumber element, and since this is clearly not a number, it
should be expected that the element will return a failure status. But even
though DecimalNumber failed to parse this token, it should not cause
the parser to halt. Instead, it should return a non-fatal fail, indicating
to the parser that it should find an alternative means to parse the token.
Specifically, it indicates to the parser that it should attempt to parse
the token with the LabelName element.
    The behavior of the DecimalNumber element in this case is an example of the
'ignore' error mode, meaning that the failure of this element should not
cause the parser to halt. This is the desired behavior in this case, but
not in others. There are cases in which DecimalNumber should cause the
parser to halt when it fails.
    Consider the token '#1az'. LC-3's grammar treats any token with a
leading '#' as a decimal number. Since '1az' is not a number, '#1az'
is an invalid token. The rule for parsing tokens of this form could
look something like the following:
    All<PoundSign, DecimalNumber>;
In english, this rule says 'if you see a pound sign, parse what comes
after it as a decimal number.' There is a problem, however. Running
this rule on the token '#1az' simply results in a non-fatal fail status
being returned. The behavior of the DecimalNumber element must be modified
to instead return a fatal fail status. And this is where the new meta
elements come in. The following change to the above rule will give the
desired behavior:
    All<PoundSign, SetError<DecimalNumber>>;
The SetError meta element sets the error mode to 'error,' meaning that
halting of the parser is allowed on parse failures. Now, when DecimalNumber
encounters '1az', will output an error to the log and return a fatal error
status.
    The utility of these new meta elements go beyond cases such as this, but
this is a nice illustrative example.