Add halting meta parser elements
These parser elements will halt the parser if their argument
element fails. This behavior is primarily useful in long Any<...>
rules. For example, consider the following rule
    Any<Number, String, Character>;             (i)
If I want the parser to halt if none of the elements succeed, I
would have to write something like
    Any<Number, String, SetError<Character>>;   (ii)
This can get rather cumbersome, and it negatively affects readability.
The HaltIfNone<...> solves this problem. The rule
    HaltIfNone<Number, String, Character>;      (iii)
is equivalent to rule (ii).