Fix validation error in hexadecimal parser
The previous iteration of the parser misidentifies the string 'x'
as a hexadecimal value. The misidentification is due to an
improper length check of the token string. For some context, the
general format for a hexadecimal value expressed as a regular
expression is /[xX][a-fA-F0-9]+/. The parser verifies that the
token begins with an 'x' and that the remaining characters are
valid hexadecimal digits. The flaw in this iteration of the parser
accepts strings of length 1, i.e. a string consisting of only 'x'.
It is vacuously true that all characters coming after the 'x' are
hexadecimal digits. In effect, the parser recognizes all strings
satisfying /[xX][a-fA-F0-9]*/ as hexadecimal values, which is
clearly distinct from the proper expression.
    The fix is quite simply to outright reject strings whose length
is less than 2.