Remove NegNumber nodes
This type of node was a hacky workaround to tack on additional
information to a DecNumber node, mainly the fact that the number
is negative. Now that I am able to store the parsed numeric value
directly within a Number node I can now perform the negation on
that parsed value.