Correctly identify branch instructions
There was a slight error in the parser that misinterpreted
any token beginning with 'br' as a branch instruction. This
was due to the code that determines if a token is or is not
the name of a valid instruction only examined the first two
characters. This meant that a label could never begin with
'br', otherwise a parse error will be thrown. This modified
check is more thorough, checking all characters of the test
string. So in other words, your labels can now start with 'br'.