Fix g++ error related to std::optional
G++ 8.1.1 throws a curious error in this function when initializing
a std::optional object with empty braces. It complained that the
code calling this function may use its return value in an
uninitialized state. While I applaud g++'s capability to detect
things like this, I think it is in the wrong here. The thing that
is truly curious about g++'s error was that it only flagged two
instances of returning empty braces out of all the others within
the same function. Furthermore, suppressing this error merely
involved invoking an alternate, but nevertheless equivalent,
form of initialization: std::nullopt. Initializing a std::optional
object with std::nullopt is documented to be equivalent to
default initialization. With this in mind, I have changed all
instances of default initialization with std::nullopt initialization
within this offending function. Doing this makes g++ happy once again.

As a side note, Clang 6.0.1 does not throw any errors in the old code.