Define .errorPrefix() method for Token objects
Sometimes I will only have token objects to work with instead of full
on AST nodes, so it is appropriate to define this method for Tokens.
The prefix format for Tokens is identical to that of SyntaxTreeNodes,
so now SyntaxTreeNode's .errorPrefix() method just redirects to its
token member's prefix method.