Replace lambda functions with static method
Several methods in the StringView class depend on a lambda that
compares the value of two string characters. The lambda functions
all have the same definitions and are defined within the methods
themselves. To reduce repetitive code, I have replaced these lambda
functions with a private static method that performs the same operation.