Move main function's code into separate function
For error conditions such as an incorrect number of command line arguments,
the original code would print to stderr and immediately terminate. This
is inconsistent with other errors, where a log summary would be printed
at the very end of the program's execution. To make the behavior of errors
like the one mentioned above consistent, I saw it appropriate to move
the code from the main function to a separate function. Now main can
detect if an error has been thrown and can display a log summary. The
code to do this will be added later.