Add simple unit testing module
The unit tests for this project have been in serious need of a
refactor. The addition of this module marks the beginning of that
process. Note that this module is still a work in progress.

I have also played with the idea of branching this module into
a separate project. Of all of the lightweight unit testing frameworks
for C++, I have not seen any that are quite this lightweight.