Warn the user of flagless branches
A branch instruction with no flags (BR) will never be taken. Hence,
it is a no-op. Usually, this is not what's wanted, so I have introduced
this warning. At the moment there is no way to suppress warnings or
errors, so if you actually intended to write such an instruction,
then you'll just have to deal with the warning. Sorry about that.