Add new unit test for StringView::operator==()
I noticed that the original unit test only tested strings with equal
sizes. I do not think that adequately tests the expected functionality.
It is much more likely in practice to test the equality of strings with
differing sizes, and I thought I should add a test for that case.