Remove explicit definitions of the "Big Five" in LC3::Value
According to C++'s "Rule-of-Zero," classes are only required to define
the "Big Five" methods in certain situations (primarily, when the purpose
of the class relates to resource ownership). Since LC3::Value does not
manage any dynamic resources, it is not only safe to remove these
definitions, but also in doing so, it improves code readability.