Make LC3::Value's constructor explicit
Consider the following code:

void f(const LC3::Value& v);

int main() {
    f(true);
}

This code is perfectly valid and will compile. The reason is because
bools can be implicitly converted to integral types, and LC3::Value's
constructor, which accepts an unsigned integral type, is not declared
"explicit." By delcaring LC3::Value's constructor "explicit," nonsense
code such as this is avoided.

Some time down the line I will be implementing some code improvements
that deal with issues such as this. This is the first step in that
effort.