Add LC3::VM::Console
The way that LC3 is structured, is presumes that input can be
obtained from the user on a character-by-character basis. In
other words, when a key is pressed, the running application consumes
it. The goal of this class is to emulate that behavior. By enabling
what is known as "raw mode", the terminal can consume keyboard input
without waiting for a newline character (which is the typical behavior).
By calling the activate() and deactivate() methods, one can enable and
disable "raw mode", respectively.

This class is designed as a singleton instance. My rationale for which
is that creating multiple Console instances can potentially cause
problems, as the state of multiple objects may become inconsistent
with each other. By restricting the class to a single instance, there
is no possibility for inconsistency (except, perhaps, in the case of
multiple threads).