Check if raw mode is active before deactivation
Without this check, there is a possibility of undefined behavior.
If one calls Console::deactivate() before any calls to Console::activate(),
the uninitialized termios struct is used to restore the old settings of the
terminal. By adding this check, it ensures that the termios struct is never
uninitialized when the settings are restored.