Unify two of CharClass' constructors
CharClass' copy constructor and its function mapping constructor
have essentially the same form. Removing CharClass' copy constructor
definition thus has no effect, so in an effort to reduce the amount
of code, I am doing just that.

Additionally, I have improved the SFINAE condition for the function
mapping constructor to something a little more robust. The previous
condition only checked the argument for the presence of a function
call operator taking in a char value, it did not, however, check
if the return value of that function call operator was a boolean.
I have modified the condition to check for that to make it clear
that it is a requirement.