Add 'seenOrigDirective' to AST analyzer flags
Because of the way that LC3 object files are structured, it makes it
difficult to support multiple '.ORIG' directives within a single
file. The 'seenOrigDirective' analyzer flag is meant to keep track
of whether the analyzer has encountered an '.ORIG' directive, and
if it encounters multiples the analyzer will throw an error.